+++ service-base/src/main/java/org/eclipse/hono/service/AbstractAdapterConfig.java
@@ -23,7 +23,6 @@ import org.eclipse.hono.client.HonoConnection;
 import org.eclipse.hono.client.RegistrationClientFactory;
 import org.eclipse.hono.client.RequestResponseClientConfigProperties;
 import org.eclipse.hono.client.TenantClientFactory;
-import org.eclipse.hono.client.impl.CommandConsumerFactoryImpl;
 import org.eclipse.hono.config.ApplicationConfigProperties;
 import org.eclipse.hono.config.ClientConfigProperties;
 import org.eclipse.hono.config.VertxProperties;
@@ -299,7 +298,7 @@ public abstract class AbstractAdapterConfig {
     @Bean
     @Scope("prototype")
     public CommandConsumerFactory commandConsumerFactory() {
-        return new CommandConsumerFactoryImpl(vertx(), commandConsumerFactoryConfig());
+        return CommandConsumerFactory.create(HonoConnection.newConnection(vertx(), commandConsumerFactoryConfig()));
     }
 
     /**